requirements.txt
blinker==1.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
proxmoxer==2.0.1
PyYAML==6.0.1
requests==2.31.0
urllib3==2.2.1
Werkzeug==3.0.2
